+++ lib/src/theme/tabbed_view_theme_constants.dart
@@ -1,6 +1,6 @@
 class TabbedViewThemeConstants {
-  static const double minimalIconSize = 8;
-  static const double defaultIconSize = 10;
+  static const double minimalIconSize = 6;
+  static const double defaultIconSize = 16;
 
   static double normalize(double buttonIconSize) {
     if (buttonIconSize >= TabbedViewThemeConstants.minimalIconSize) {